move buildRasRedirectUrl() to access-utils
